The v1 implementation of the partial init code
This iteration of the partial init design now includes explicit errors
for all possible (non timeout) failures. It also makes it easier for the
caller to determine which components failed if a chip is partially
initialized.

As a part of this change the druken_monkey test was rewritten to be run
in a CI environment.