v2 implementation of the partial init overhaul
In previous iterations of the change all components were treated
independendly from eachother. Meaning that the initialization and user
updates were go through ARC, ETH, and then DRAM. Now all are initialized
simultaniously and the user is updated on each as they are polled.

This change also allows for fewer generics to be passed around as the
code which monitors the current state of the initialization no longer
has to be generic accross each component as it recieves status updates
from all components every tick.