Populated the code to detect DRAM initialization
Previously the skeleton of the DRAM init code existed in the init
changes. However it was assumed that if ARC initialized then so would
DRAM, and this is true to an extent. However there are cases in which
some DRAM channels won't train. This change now explicity checks if any
DRAM channels are not trained and will no longer incorrectly consider a
chip to be initialized if any are not.