Handle the case where a chip has been wiped
This should never happen in practice, but this code change makes sure
that a chip which does not have any fw loaded will be correctly reported
as such. It also improves the output message to the user to make it
clear that un uninitialized ARC will mean that the status of DRAM and
ETH will not be checked.