Allow chip initialization to deduplicate detected chips
After overhauling the chip initialization routine so that it is faster,
more reliable and less likely to cause a chip in a bad state to
completely fail the chip detection routine was overcounting the number
of detected chips.

This is not a problem in itself because the chips were still correctly
returned. However the monitoring function did end up reporting the
incorrect number of chips while the detection process was ongoing. To
work around this problem I added a signal to the monitoring process to
tell it to correct down and thereby remove a duplicated chip from the
ongoing count.