Added update callback to pyluwen detect_chips
This adds the ability to specify a callback for the detect_chips
function. The callback accepts a single argument which is the current
update status of the chip. This should be considered the MVP
implementation of this feature but is being pushed now because when used
it allows ctrl-c to interupt the init process. Currently the init
process is not interuptable until after the scan has completed.

This is not a breaking change, the default behaviour is to be silent.
However ctrl-c will still work.