Plummed python errors from callback to init function
In the previous commit I left any error as a result of the callback
function to panic (via unwrap). This is extremely bad from a python
perspective because we can no longer perform any error handling if a
ctrl-c is hit, or a legitimate error is thrown from Python code. To
resolve this I added a -> Result<(), E> to the callback function and
plummed it through the code back to the calling init function so that it
could eventually be returned as a thrown Exception in the Python code.
For Rust the default behaviour is to print that an exception happened
and continue.