luwen-cem: Added boardtype attribute to the chips
This adds a new "boardtype" dict to the output of cem. This is a first
pass implementation that has a few limitations, 1 both chips in an n300
will be called out as being n300 without differentiating between left
and right. 2 old boards will get the "UNSUPPORTED" value key instead of
None, or in the case of yaml null.