Fixed wh/gs bcast to correctly select the paramters for each chip
Previously the broadcast parameters were hardcoded to the wormhole chip
definition. This commit updates the pci interface module to select the
broadcast parameters based on the device architecture. For now this is a
selection between Wormhole and Grayskull, but will need to be updated to
include Blackhole in the future.