Only check for the telemetry address once
This is a quick fix for the issue where we may hang smi when running a
workload alongside it. Note that this still will mean that if you start
smi after you start a workload there will be around a 100ms window where
you could trigger this issue.

A proper fix for gs would involve using the driver resource locks (available in
1.27) to avoid any contention on the arc. WH is a bit easier because we
can use the hardware semaphore + mutex registers to solve the issue.
However both require by in from umd before they fully solve the problem.