Updated Grayskull chip init to allow pre 1.0.0 fw
This change is in response to an issue where dram init would fail with
an error if the firmware was too old to retrieve the current fw version
via an arc message. Now instead of aborting it will correctly report
that the arc fw is too old (but not provide the current version in the
error messsage).