Fixed bug where luwen would think dma was happening
When reading a single bit from the top of a value the old read_field
implementation would not zero out the remaining bits of the returned byte.
This commit fixes this bug (adding a test to catch it in the future) and
cleans up the read_modify implementation so it will be easier to maintin.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>