Moving debian to debian.debian
Why?  Ok I want to keep the build stuff next to the respective
repository / as easy to track as possible.

Problem: Ubuntu and Debian have utterly diverged for now on how to build
/ deal with Rust dependencies / building stuff.  So the build processes
are, unfortunately, different and obnoxious.

Solution: ok track things differently for now and hope it sorts out at
some point.

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>