Added support for BH style broadcast
The BH broadcast when cooridnates are translated requiring different
coordiantes to be programmed and not allowing backwards compatibility
with untranslated coordinates breaks some fundimental assumptions backed
in the luwen's design. Instead of rewriting these assumptions I opted to
support this broadcast for the one case I currently care about (burnin)
and develop a more comprehensive redeisgn in a seperate branch.

Signed-off-by: Daniel Rosen <3820236+DRvader@users.noreply.github.com>