fix: fixed telemetry test
The telemetry test was originally written for WH, however it is
also important to get coverage for BH. So I rewrote some of the
telemetry gathering logic to unify WH/BH telemetry data for a subset of
useful values. Avoided changing too much of the data API because that
should be done in a larger API redesign to avoid too much churn.

Signed-off-by: Daniel Rosen <3820236+DRvader@users.noreply.github.com>