support telemetry buffer offsets > 256 bytes for blackhole
blackhole telemetry parsing code casts the telemetry buffer index to a
u8, which means that the effective offset of telemetry data is limited
to 256 bytes. This means we cannot use an offset greater than 62, as the
slicing code will fail with index 63 (63 * 4 + 4 = 256). Use a u16 for
the offset to resolve this, and use a u16 for the tag count as well for
future expansion.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>