blackhole protobufs: trim 0s after the padding from proto_bin
In remove_padding_proto_bin, remove four trailing 0s from the proto_bin if
present before removing the conventional padding. This only applies if the
proto_bin is padded to a multiple of 8 bytes.

The proto_bin padding should be one of (0), (0,1), (0,1,2), or (0,1,2,3).
Protobuf encoding on the FW side handles padding to a multiple of 4 bytes using
this padding format, but FW versions 80.18.0 - 18.5.0 further pad each boot_fs
section to a multiple of 8 bytes using extra 0s.

This fixes the decode_boot_fs_table function.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>