Convert format strings to make clippy 1.88 happy
Newer versions of clippy prefers that the variables printed by
format!() and friends are inline as part of the format string instead
of arguments. Adjust the code accordingly.

This was added in rust 1.58, in 2022. So I think we're safe to assume
that we'll always have a newer compiler than that (current dependencies
seem to require 1.82 or newer).

Signed-off-by: Olof Johansson <olof@lixom.net>