Replace vendored rust-protoc with checked-in generated protobuf code
This code doesn't change often, and it uses a questionable approach of
pulling in a personal repo as a dependency.

Given that the generated files change so rarely, instead go with checked
in versions, and include instructions on how to regenerate them.

This also removes a ton of transitive dependencies from Cargo.lock,
and removes the need of a custom build.rs.