Remove Debian 12 pre-cursor work
Debian 12 had some pre-cursor work put in early on, before we ran quite
into the problem of how much of the rust stack was missing in Debian 12.
Ultimately Debian 12 isn't going to get fully supported the way we would
like, but Debian 13 looks like it will be fine (see prior commit).

This cleans up what's left of the early stages on Debian 12 and leaves
us as the minimum Debian version as 13

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>