BH telem: check if ARC not ready before checking if address out of bounds
Currently there is a check first to see if the telem_struct_addr is out of
bounds of CSM memory. This covers if telem_struct_addr is 0, and will throw a
more generic error than ArcNotReady.

Check if telem_struct_addr is 0 before checking if it's out of bounds to more
clearly indicate that ARC is not ready.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>