build(cargo): set MSRV to 1.75.0 (#85)
It seems that the most recent Rust feature the codebase depends on is
`pointer_byte_offsets`, introduced in v1.75.0. Conveniently, this is
also the precise version provided on Ubuntu 22.04, which we currently
support. As such, it makes sense to explicitly codify that requirement
by adopting a policy for a Minimum Supported Rust Version (MSRV).

This does not prevent using a newer version of the toolchain (as is
guaranteed by Rust[^1]).

Enforcing this policy within the `Cargo.toml` also allows us to fix the
version of the committed `Cargo.lock` for free, which ensures it will
not be bumped automatically by newer toolchains.[^2]

References:

[^1]: https://blog.rust-lang.org/2014/10/30/Stability/
[^2]: rust-lang/cargo#12861