feat(demo): move to examples
Since this binary is a demonstration of how to use `luwen`'s API, it is
better suited as an example, where it is only compiled when the user
explicitly wants to build examples. This is also the idiomatic (Rusty)
place for demo binaries to be placed, making it more discoverable.

Signed-off-by: Zakhary Kaplan <zkaplan@tenstorrent.com>