build(cargo): share version numbering
Since the organisational refactor, it no longer makes sense to have
separate versions for each sub-crate, as the intended interface for
consumers is through the top-level `luwen` crate.

This commit also extends version sharing to bindings (notably
`pyluwen`). Semantically, this forwards breaking changes from Rust to
the Python interface. This may be decoupled in a future commit if found
necessary.

Signed-off-by: Zakhary Kaplan <zkaplan@tenstorrent.com>