fix: repair response data in `bh_arc_msg`
Issue was caused by the updated semantics for the return value of
`bh_arc_msg` not bring properly updated at the call sites.

The return value was updated to return the response buffer bytes `[0:7]`
rather than `[1:7]`, which means call sites using this value need to
have been updated as well.

Signed-off-by: Zakhary Kaplan <zkaplan@tenstorrent.com>